% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det_model_fxn.r
\name{det_model}
\alias{det_model}
\title{Deterministic model function}
\usage{
det_model(params)
}
\arguments{
\item{params}{A list with the following parameters included: 

ini.prev = percentage of fawns infected at the start (scalar value between 0 and 1),  

n.age.cats = number of age categories to monitor. Currently makes the most sense to be by decade for 10 categories (must be scalar value greater than 3).

n.i.cats = number of I subcategories

n.e.cats = number of E subcategories

e.move = rate of movement in the exposed categories (scalar values between 0 and 1). 

i.move = rate of movement in the infectious categories (scalar between 0 and 1)

beta = transmission coefficient   

theta = effect of population size on transmission (1 = frequency dependence, 0 = density dependent).  

contacts = symmetrical matrix of n.age.cats*n.age.cats of the contact rate between age categories. 1 = average, 1.1 = 10% increase, .9 = 10% decrease. 

n0 = initial population size (scalar value greater than 0)

n.days = number of days to run the model (scalar value greater than 2),

severity is a vector of length n.age.cats of the proportion of I that goes to the hospital

d.no is the proportion of severe cases that die per day w/o hospital care (currently scalar)

d.yes is the proportion that die with hospital care

beds is the # of beds or ICUs that are available}
}
\value{
A list with 3 outputs: 

1. List of counts of the # of individuals in S, E, I, R categories over time with the following columns: age category, day of simulation, sub category (for E and I), n = number of individuals
  
2. List of outcomes-- Dt = deaths, Ht = hospitalizations needed

 Lists for Dt and Ht with the following columns: age category, day of the simulation, n = # of individuals 
 
3. R0 = basic disease reproductive number
}
\description{
Deterministic daily age structured model
}
\examples{
n.age.cats <- 10

params <- list(ini.prev = 0.02, n.age.cats = n.age.cats,  n.e.cats = 10, 
n.i.cats = 10, e.move = 0.4, i.move = 0.4, beta = 0.08, theta = 1, 
n0 = 10000, n.days = 360, contacts = matrix(1, nrow = n.age.cats, ncol = n.age.cats),
severity = seq(0.01, 0.2, length.out = n.age.cats), d.no = 0.05, d.yes = 0.01,
beds = 100)

out <- det_model(params)

}
