% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{launches the shinyApp}
\usage{
launchApp(example)
}
\arguments{
\item{example}{name of the application to run. Current apps are only "app"}
}
\description{
launches the shinyApp
}
\examples{
launchApp("app") 

}
