% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_dist.r
\name{plot_age_dist}
\alias{plot_age_dist}
\title{Plot the age distribution}
\usage{
plot_age_dist(dat)
}
\arguments{
\item{dat}{dataframe of counts provided as output from the COVID model 
functions}
}
\value{
a plot the age distribution at the end point. Note that the final 
age category includes all individuals of that age or more
}
\description{
Plot the age distribution
}
\examples{
n.age.cats <- 10

params <- list(ini.prev = 0.02, n.age.cats = n.age.cats,  n.e.cats = 10, 
n.i.cats = 10, e.move = 0.4, i.move = 0.4, beta = 0.08, theta = 1, n0 = 10000, 
n.days = 360, contacts = matrix(1, nrow = n.age.cats, ncol = n.age.cats),
severity = seq(0.01, 0.2, length.out = n.age.cats), d.no = 0.1, d.yes = 0.02,
beds = 20)

out <- det_model(params)
plot_age_dist(out$counts)

}
