% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contacts.r
\name{plot_contacts}
\alias{plot_contacts}
\title{Plot the contact matrix}
\usage{
plot_contacts(dat)
}
\arguments{
\item{contacts}{a symmetrical matrix of contact rates by age}
}
\value{
a heatmap of that contact matrix.
}
\description{
Plot the contact matrix
}
\examples{
plot_contacts(contacts)

}
