% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exposure_age_end.r
\name{plot_exposed_age_end}
\alias{plot_exposed_age_end}
\title{Plot the exposure by age at the end of the simulation}
\usage{
plot_exposed_age_end(dat)
}
\arguments{
\item{dat}{counts provided as output from the COVID model function}
}
\value{
a plot of the prevalence by age
}
\description{
Plot the exposure by age at the end of the simulation
}
\examples{
n.age.cats <- 10

params <- list(ini.prev = 0.02, n.age.cats = n.age.cats,  n.e.cats = 10, 
n.i.cats = 10, e.move = 0.4, i.move = 0.4, beta = 0.08, theta = 1, 
n0 = 10000, n.days = 360, contacts = matrix(1, nrow = n.age.cats, ncol = n.age.cats),
severity = seq(0.01, 0.2, length.out = n.age.cats), d.no = 0.05, d.yes = 0.01,
beds = 100)

out <- det_model(params)
plot_exposed_age_end(out$counts)

}
