% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_periods.r
\name{plot_periods}
\alias{plot_periods}
\title{Plot time in exposed and infectious categories}
\usage{
plot_periods(e.move, i.move, n.e.cats, n.i.cats)
}
\arguments{
\item{e.move}{proportion or probability of progressing through the E 
category. Must be between 0 and 1.}

\item{i.move}{proportion or probability of progressing through the I
category. Must be between 0 and 1.}

\item{n.e.cats}{number of exposed sub categories}

\item{n.i.cats}{number of infectious sub categories}
}
\value{
a density plot of time in the E or I category.
}
\description{
Creates a plot of the time (in days) a person spends in exposed and infectious
categories. Uses 1000 draws from a Gamma distribution.
}
\examples{
plot_periods(e.move = 0.95, i.move = 0.95, n.e.cats = 5, n.i.cats = 3)

}
