% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prev_time.r
\name{plot_prev_time}
\alias{plot_prev_time}
\title{Plot the prevalence over time}
\usage{
plot_prev_time(dat)
}
\arguments{
\item{dat}{list of counts provided as output from the det model functions}
}
\value{
a plot of the prevalence over time
}
\description{
Plot the prevalence over time
}
\examples{
n.age.cats <- 10

params <- list(ini.prev = 0.02, n.age.cats = n.age.cats,  n.e.cats = 10, 
n.i.cats = 10, e.move = 0.4, i.move = 0.4, beta = 0.08, theta = 1, 
n0 = 10000, n.days = 360, contacts = matrix(1, nrow = n.age.cats, ncol = n.age.cats),
severity = seq(0.01, 0.2, length.out = n.age.cats), d.no = 0.05, d.yes = 0.01,
beds = 100)

out <- det_model(params)
plot_prev_time(out$counts)
}
