% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_severity.r
\name{plot_severity}
\alias{plot_severity}
\title{Plot the severity versus age}
\usage{
plot_severity(s.lo, s.hi, n.age.cats)
}
\arguments{
\item{s.lo}{proportion of severe cases in the lowest age category}

\item{s.hi}{proportion of severe cases in the highest age category}

\item{n.age.cats}{number of age categories}
}
\value{
a plot of severity versus age
}
\description{
Creates a plot of the proportion of severe cases by age category. Severe is 
assumed to be those that would need hospitalization
}
\examples{
plot_severity(s.lo = 0.01, s.hi = 0.2, n.age.cats = 10)

}
